moved delete and show methods to timeslots controller
