+++ app/Http/routes.php
@@ -16,16 +16,20 @@ Route::group(['middleware' => 'web'], function () {
             name('presentation.decline');
         Route::post('{id}/decline', 'PresentationsController@save_comment')->
             name('presentation.comment');
-        Route::get('schedule/{display_room?}', 'PresentationsController@show_schedule')->
-            name('presentation.schedule');
         Route::post('updateSchedule/{display_room?}', 'PresentationsController@update_schedule')->
             name('presentation.book');
-        Route::get('deleteTime/{display_room}/{id}', 'PresentationsController@deleteTime')->
-            name('delete_time');
         Route::post('addTime/{display_room}', 'PresentationsController@addTime')->
             name('add_time');
         Route::get('/{status}', 'PresentationsController@index')->name('presentation.status');
     });
+
+    Route::group(['prefix' => 'timeslot'], function (){
+        Route::get('deleteTime/{display_room}/{id}', 'TimeslotController@deleteTime')->
+            name('delete_time');
+        Route::get('schedule/{display_room?}', 'TimeslotController@show_schedule')->
+            name('timeslot.show');
+    });
+
 
     Route::group(['prefix' => 'conference'], function (){
         Route::get('/old', 'ConferencesController@old')->
