+++ resources/views/user/admin/_sidebar.blade.php
@@ -1,6 +1,6 @@
 @if(current_conference())
 <ul class="nav nav-pills nav-stacked"> Current Conference
-    <li><a href="{{ route('presentation.schedule')}}">Schedule</a></li>
+    <li><a href="{{ route('timeslot.show')}}">Schedule</a></li>
     <li><a href="{{ route('presentation.index') }}">Presentations</a></li>
 </ul>
 
@@ -17,5 +17,3 @@
     <li><a href="{{ route('room.index') }}">Rooms</a></li>
     <li><a href="{{ route('conference.create') }}">Create new Conference</a></li>
 </ul>
-
-
